% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Clean_data
Load downloaded data, clean the dataset, strip extra white spaces from column names, convert column names to lower case, remove the NA rows and return clean data}
\usage{
clean_data(data)
}
\arguments{
\item{data.frame}{Data set to clean}
}
\value{
data A cleaned and simplified DataFrame of the relevant columns for summary and visualization
}
\description{
Clean_data
Load downloaded data, clean the dataset, strip extra white spaces from column names, convert column names to lower case, remove the NA rows and return clean data
}
\examples{
df <- clean_data(data)
}
