% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_classifier.R
\name{fit_classifier}
\alias{fit_classifier}
\title{Fit dummy classifier and Logistic regression models}
\usage{
fit_classifier(
  train_df,
  target_col,
  numeric_feats = NULL,
  categorical_feats = NULL,
  cv = 5
)
}
\arguments{
\item{train_df}{dataframe that will be used to train the model}

\item{target_col}{The column that needs to be classified as a string}

\item{numeric_feats}{The numeric columns as a list}

\item{cv}{The number of cross validation folds as an integer}

\item{categorical_features}{The categorical columns as a list}
}
\value{
A data frame
}
\description{
Fit dummy classifier and Logistic regression models
}
\examples{
library(tidyverse)
fit_classifier(data, target_col = 'income', categorical_feats=c('occupation'))
fit_classifier(data, target_col = 'income', numeric_feats = c('age', 'fnlwgt'), categorical_feats=c('genre'), cv=10)

}
