% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_regressor.R
\name{fit_regressor}
\alias{fit_regressor}
\title{Fit dummy regressor and linear regression models}
\usage{
fit_regressor(
  train_df,
  target_col = NULL,
  numeric_feats = NULL,
  categorical_feats = NULL,
  cv = 5
)
}
\arguments{
\item{train_df}{dataframe that will be used to train the model}

\item{target_col}{The column that needs to be classified as a string}

\item{numeric_feats}{The numeric columns as a vector character}

\item{categorical_feats}{The categorical columns as a vector character}

\item{cv}{The number of cross validation folds as an integer}
}
\value{
A data frame
}
\description{
Fit dummy regressor and linear regression models
}
\examples{
fit_regressor(gapminder::gapminder, target_col="gdpPercap", numeric_feats=c("pop"), categorical_feats <- c("continent"), cv =5)
fit_regressor(gapminder::gapminder, target_col="gdpPercap", numeric_feats=c("year", "lifeExp", "pop"), categorical_feats <- c("continent"), cv =5)
}
