% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corr.R
\name{plot_corr}
\alias{plot_corr}
\title{Returns correlation plot for all the columns in dataset}
\usage{
plot_corr(data, corr = "pearson", pair_cols = NULL)
}
\arguments{
\item{data}{a dataframe}

\item{corr}{type of correlation: 'spearman', 'kendall' or 'pearson'}

\item{pair_cols}{: a character vector od columns for which you want correlation plot}
}
\value{
corr_plot, a ggplot chart object
}
\description{
Returns correlation plot for all the columns in dataset
}
\examples{
data <- data.frame(age = c(1, 2, 4, 6, 8),education.num = c(3, 6, 12, 18, 24), target=c(20, 30, 12, 0, 2))
plot_corr(data, corr = 'pearson', pair_cols=c('age', 'education.num'))
plot_corr(data)
}
