% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Clean and return dataframe
Loads dataframe, removes Nan rows, strips whitespaces, converts columns to lowercase and returns a clean dataframe}
\usage{
clean_data(data)
}
\arguments{
\item{data}{: A raw dataframe}
}
\value{
data : clean dataframe
}
\description{
Clean and return dataframe
Loads dataframe, removes Nan rows, strips whitespaces, converts columns to lowercase and returns a clean dataframe
}
\examples{
library(tidyverse, quietly = TRUE)
data= tibble(num_legs= c(2, 4, 8, 0), num_wings= c('bad', 'Good', NA, 'bad'),Num_specimen = c(10, 2, 1, 8))
}
