% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_classifier.R
\name{fit_classifier}
\alias{fit_classifier}
\title{Fit dummy classifier and Logistic regression models}
\usage{
fit_classifier(
  train_df,
  target_col,
  numeric_feats = list(),
  categorical_feats = list(),
  cv = 5
)
}
\arguments{
\item{train_df}{dataframe that will be used to train the model}

\item{target_col}{The column that needs to be classified as a string}

\item{numeric_feats}{The numeric columns as a list}

\item{categorical_feats}{The categorical columns as a list}

\item{cv}{The number of cross validation folds as an integer}
}
\value{
A data frame
}
\description{
Fit dummy classifier and Logistic regression models
}
\examples{
library(tidyverse)
library(lattice)
library(caret)
library(mltools)
library(data.table)
data <- data.frame(x = c(1, 2, 4, 6, 8), y = c("Yes", "No", "Maybe", "Someday", "Never"), z = c(0, 1, 1, 0, 1))
fit_classifier(data, target_col = 'z', numeric_feats = list(), categorical_feats = list('y'), cv = 10)
fit_classifier(data, target_col = 'z', numeric_feats = list('x'), categorical_feats = list('y'), cv = 5)
}
