% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_classifier.R
\name{fit_classifier}
\alias{fit_classifier}
\title{Fit dummy classifier and Logistic regression models}
\usage{
fit_classifier(
  train_df,
  target_col,
  numeric_feats = NULL,
  categorical_feats = NULL,
  cv = 5
)
}
\arguments{
\item{train_df}{dataframe that will be used to train the model}

\item{target_col}{The column that needs to be classified as a string}

\item{numeric_feats}{The numeric columns as a list}

\item{categorical_feats}{The categorical columns as a list}

\item{cv}{The number of cross validation folds as an integer}
}
\value{
A data frame
}
\description{
Fit dummy classifier and Logistic regression models
}
\examples{
library(tidyverse)
library(lattice)
library(caret)
library(mltools)
library(data.table)
library(devtools)
devtools::install_github("jennybc/gapminder")
library(gapminder)
gapminder <- gapminder |> dplyr::filter(continent=="Asia" | continent=="Europe")
gapminder$country <- as.character(gapminder$country)
gapminder$continent <- as.character(gapminder$continent)
gapminder$year <- as.numeric(gapminder$year)
gapminder$pop <- as.numeric(gapminder$pop)
fit_classifier(gapminder, target_col = 'continent', numeric_feats = list('gdpPercap'), categorical_feats = list('country'), cv = 5)
}
